# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FacturacionDemoAGVS API",
    "description": "API para la gestión de documentos y facturación.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local de desarrollo"
    }
  ],
  "paths": {
...
    "/producto": {
      "get": {
        "tags": [
          "producto"
        ],
        "summary": "Obtener todos los productos",
        "description": "Devuelve una lista de todos los productos registrados en el sistema.",
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "descripcion": {
                        "type": "string",
                        "example": "Producto de alta calidad"
                      },
                      "precio": {
                        "type": "number",
                        "example": 50.0
                      },
                      "status": {
                        "type": "string",
                        "example": "activo"
                      },
                      "stock": {
                        "type": "integer",
                        "example": 100
                      },
                      "exento": {
                        "type": "boolean",
                        "example": false
                      },
                      "descuento": {
                        "type": "number",
                        "example": 10.0
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "producto"
        ],
        "summary": "Crear o recuperar un producto",
        "description": "Crea un nuevo producto o recupera uno existente basado en los datos proporcionados.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "descripcion": {
                    "type": "string",
                    "example": "Producto de alta calidad"
                  },
                  "precio": {
                    "type": "number",
                    "example": 50.0
                  },
                  "status": {
                    "type": "string",
                    "example": "activo"
                  },
                  "stock": {
                    "type": "integer",
                    "example": 100
                  },
                  "exento": {
                    "type": "boolean",
                    "example": false
                  },
                  "descuento": {
                    "type": "number",
                    "example": 10.0
                  }
                },
                "required": [
                  "descripcion",
                  "precio",
                  "status",
                  "stock",
                  "exento"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado o recuperado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "Producto de alta calidad"
                    },
                    "precio": {
                      "type": "number",
                      "example": 50.0
                    },
                    "status": {
                      "type": "string",
                      "example": "activo"
                    },
                    "stock": {
                      "type": "integer",
                      "example": 100
                    },
                    "exento": {
                      "type": "boolean",
                      "example": false
                    },
                    "descuento": {
                      "type": "number",
                      "example": 10.0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos proporcionados"
          }
        }
      }
    },
    "/producto/{producto_id}": {
      "get": {
        "tags": [
          "producto"
        ],
        "summary": "Obtener un producto por ID",
        "description": "Devuelve los detalles de un producto específico identificado por su ID.",
        "parameters": [
          {
            "name": "producto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "Producto de alta calidad"
                    },
                    "precio": {
                      "type": "number",
                      "example": 50.0
                    },
                    "status": {
                      "type": "string",
                      "example": "activo"
                    },
                    "stock": {
                      "type": "integer",
                      "example": 100
                    },
                    "exento": {
                      "type": "boolean",
                      "example": false
                    },
                    "descuento": {
                      "type": "number",
                      "example": 10.0
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "producto"
        ],
        "summary": "Actualizar un producto por ID",
        "description": "Actualiza los detalles de un producto específico identificado por su ID.",
        "parameters": [
          {
            "name": "producto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "descripcion": {
                    "type": "string",
                    "example": "Producto Actualizado"
                  },
                  "precio": {
                    "type": "number",
                    "example": 60.0
                  },
                  "status": {
                    "type": "string",
                    "example": "inactivo"
                  },
                  "stock": {
                    "type": "integer",
                    "example": 50
                  },
                  "exento": {
                    "type": "boolean",
                    "example": true
                  },
                  "descuento": {
                    "type": "number",
                    "example": 5.0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "Producto Actualizado"
                    },
                    "precio": {
                      "type": "number",
                      "example": 60.0
                    },
                    "status": {
                      "type": "string",
                      "example": "inactivo"
                    },
                    "stock": {
                      "type": "integer",
                      "example": 50
                    },
                    "exento": {
                      "type": "boolean",
                      "example": true
                    },
                    "descuento": {
                      "type": "number",
                      "example": 5.0
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "400": {
            "description": "Datos inválidos proporcionados"
          }
        }
      }
    },
    "/producto/codigo/{codigo}": {
      "get": {
        "tags": [
          "producto"
        ],
        "summary": "Obtener un producto por código",
        "description": "Devuelve los detalles de un producto específico identificado por su código.",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "PROD123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/producto/codigo_barras/{codigo_barras}": {
      "get": {
        "tags": [
          "producto"
        ],
        "summary": "Obtener un producto por código de barras",
        "description": "Devuelve los detalles de un producto específico identificado por su código de barras.",
        "parameters": [
          {
            "name": "codigo_barras",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1234567890123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/producto/codigo_QR/{codigo_QR}": {
      "get": {
        "tags": [
          "producto"
        ],
        "summary": "Obtener un producto por código QR",
        "description": "Devuelve los detalles de un producto específico identificado por su código QR.",
        "parameters": [
          {
            "name": "codigo_QR",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "QR123456789"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
...
  }
}